refactor(insights-ui): single cron heartbeat for all report generation
Collapses the three ~3-5 min crons (generate-ticker-v1-request,
generate-etf-v1-request, auto-generation tick) into one generic
/cron/heartbeat invoked every 3 min. Each tick enqueues any due
auto-generation batches, then advances pending stock and ETF requests.

- Extract processor logic into reusable functions (processPendingTickerRequests,
  processPendingEtfRequests); the existing generate-*-v1-request routes become
  thin wrappers over them (so manual runs + the etf wait-for-generation script
  keep working).
- Add runKoalaGainsHeartbeat: runs enqueue + both processors, each isolated in
  its own try/catch so one failure never blocks the others; enqueue runs first
  so a fresh batch can start processing in the same tick.
- Remove the standalone /auto-generation/tick route (folded into the heartbeat);
  runAutoGenerationTick stays as the enqueue step.
- scheduler.tf: three cron entries become one koala_gains_heartbeat
  (rate(3 minutes)). daily gainers/losers crons untouched.
