+++ insights-ui/src/utils/auto-generation/auto-etf-generation-utils.ts
@@ -22,10 +22,9 @@ import { ensureMorDataForAnalysis } from '@/utils/etf-analysis-reports/mor-scrap
  * reuses the same gates. The difference is the selection: instead of the OLDEST
  * already-generated reports, this picks ETFs whose reports have NOT been
  * generated/tried yet, prioritized by country — US first, then Canada, then
- * everything else. Called every few minutes by the `/auto-generation/tick`
- * dispatcher (and by the `enqueue-auto-etf-generation` route for manual runs); the
- * existing ~3-min `generate-etf-v1-request` processor then generates whatever this
- * enqueues.
+ * everything else. Called every few minutes by the `/cron/heartbeat` job (and by
+ * the `enqueue-auto-etf-generation` route for manual runs); the same heartbeat's
+ * processor step then generates whatever this enqueues.
  */
 
 const US_EXCHANGES: string[] = Object.values(EtfUSExchanges);
