+++ insights-ui/src/utils/auto-generation/auto-stock-generation-utils.ts
@@ -17,7 +17,7 @@ import { getOldestStocksOverall } from '@/utils/oldest-reports-utils';
 /**
  * Claude-usage-gated stock report auto-generation (enqueue side).
  *
- * Called every few minutes by the `/auto-generation/tick` dispatcher (and by the
+ * Called every few minutes by the `/cron/heartbeat` job (and by the
  * `enqueue-auto-stock-generation` route for manual runs). Each call is gated by the
  * App Settings controls: the GENERATION_ENTITY must include stocks, the current ET
  * time must fall inside the selected GENERATION_WINDOW, no auto batch may be open,
